requirements.txt
<<<<<<< HEAD
flask
flask_sqlalchemy
pandas
=======
flask
flask_sqlalchemy
pandas
matplotlib
ofxparse
PyMuPDF
pytesseract
Pillow
>>>>>>> Atualização do app.py com suporte a data
